[ Feat ] - Criando rota Supervisor realizando ajustes nas demais rotas

+++ src/main/java/com/logimarui/core/api/domain/read/Produto.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -11,5 +13,5 @@ public class Produto {
    private Long codigo;
    private String descricao;
    private String embalagem;
-   private Float peso;
+   private BigDecimal peso;
 }
