+++ src/main/java/com/logimarui/core/api/dto/produto/ProdutoResponseDTO.java
@@ -1,8 +1,10 @@
 package com.logimarui.core.api.dto.produto;
 
+import java.math.BigDecimal;
+
 public record ProdutoResponseDTO (
         Long codigo,
         String descricao,
         String embalagem,
-        Float peso
+        BigDecimal peso
 ){}
